--- Snippets/Core/Core_7.8/Recoverability/Delayed/CustomPolicies/MyBusinessException.cs
@@ -1,9 +1,0 @@
-﻿namespace Core7.Recoverability.Delayed.CustomPolicies
-{
-    using System;
-
-    class MyBusinessException :
-        Exception
-    {
-    }
-}
