+++ Snippets/Core/Core_7/Container/ExternallyManaged.cs
