+++ Snippets/Core/Core_7/Conventions/AttributeMessageConvention.cs
