+++ Snippets/Core/Core_7/Conversation/NewConversation.cs
