+++ Snippets/Core/Core_7/Notifications/SubscribeToNotifications.cs
@@ -1,65 +1,37 @@
-﻿
+﻿using System.Threading.Tasks;
 
 namespace Core7.Notifications
 {
-    using System.Text;
     using NServiceBus;
-    using NServiceBus.Faults;
     using NServiceBus.Logging;
 
     class SubscribeToNotifications
     {
         static ILog log = LogManager.GetLogger<SubscribeToNotifications>();
 
-        #pragma warning disable 618
-        #region SubscribeToErrorsNotifications
-
-        void EndpointStartup()
+        void EndpointStartup(EndpointConfiguration endpointConfiguration)
         {
-            var endpointConfiguration = new EndpointConfiguration("EndpointName");
-            Subscribe(endpointConfiguration.Notifications);
+            #region SubscribeToErrorsNotifications
+            var recoverability = endpointConfiguration.Recoverability();
+
+            recoverability.Immediate(settings => settings.OnMessageBeingRetried(retry =>
+            {
+                log.Info($"Message {retry.MessageId} will be retried immediately.");
+                return Task.CompletedTask;
+            }));
+
+            recoverability.Delayed(settings => settings.OnMessageBeingRetried(retry =>
+            {
+                log.Info($@"Message {retry.MessageId} will be retried after a delay.");
+                return Task.CompletedTask;
+            }));
+
+            recoverability.Failed(settings => settings.OnMessageSentToErrorQueue(failed =>
+            {
+                log.Fatal($@"Message {failed.MessageId} will be sent to the error queue.");
+                return Task.CompletedTask;
+            }));
+            #endregion
         }
-
-        void Subscribe(Notifications notifications)
-        {
-            var errors = notifications.Errors;
-
-            errors.MessageHasBeenSentToDelayedRetries += LogEvent;
-
-            errors.MessageHasFailedAnImmediateRetryAttempt += LogEvent;
-            errors.MessageSentToErrorQueue += LogEvent;
-        }
-
-        static string GetMessageString(byte[] body)
-        {
-            return Encoding.UTF8.GetString(body);
-        }
-
-        void LogEvent(object sender, FailedMessage failed)
-        {
-            log.Fatal($@"Message sent to error queue.
-        Body:
-        {GetMessageString(failed.Body)}");
-        }
-
-        void LogEvent(object sender, DelayedRetryMessage retry)
-        {
-            log.Info($@"Message sent to Delayed Retries.
-        RetryAttempt:{retry.RetryAttempt}
-        Body:
-        {GetMessageString(retry.Body)}");
-        }
-
-        void LogEvent(object sender, ImmediateRetryMessage retry)
-        {
-            log.Info($@"Message sent to Immedediate Retry.
-        RetryAttempt:{retry.RetryAttempt}
-        Body:
-        {GetMessageString(retry.Body)}");
-        }
-
-        #endregion
-        #pragma warning restore 618
     }
-
 }
