+++ Snippets/Core/Core_7/PubSub/Publishing/DisablePublishing.cs
