+++ Snippets/Core/Core_7/Recoverability/ConsecutiveFailures.cs
