+++ Snippets/Core/Core_7/Sagas/FindByHeader/MyMessage.cs
