+++ Snippets/Core/Core_7/Sagas/FindByHeader/MySaga.cs
