+++ Snippets/Core/Core_7/Sagas/FindByHeader/MySagaData.cs
