+++ Snippets/Core/Core_7/Sagas/ReverseMappingApi/MyFirstMessage.cs
