+++ Snippets/Core/Core_7/Sagas/ReverseMappingApi/MySaga.cs
