+++ Snippets/Core/Core_7/Sagas/ReverseMappingApi/MySagaData.cs
