+++ Snippets/Core/Core_7/Sagas/ReverseMappingApi/MySecondMessage.cs
