+++ Snippets/Core/Core_7/Sagas/SagaAnalyzers.cs
