+++ Snippets/Core/Core_7/Scanning/DisableAssemblyFileScanning.cs
