+++ Snippets/Core/Core_7/Serialization/DisableContentTypeInference.cs
