+++ Snippets/Core/Core_7/Serialization/DisableDynamicTypeLoading.cs
