--- nservicebus/recoverability/configure-error-handling_errorqueuedefault_core_[6.3,).partial.md
@@ -1,1 +1,0 @@
-The default error queue name is `error` but some transports require it to be explicitly configured.
